Remove CreateInstanceFunc that used reflection to create constructor and used source generator instead.
