--- CommonValueObjects/generator/CamelCaseExtension.cs
@@ -1,15 +1,0 @@
-﻿namespace FunctionalDDD.CommonValueObjectGenerator
-{
-
-    public static class CamelCaseExtension
-    {
-        public static string ToCamelCase(this string str)
-        {
-            if (!string.IsNullOrEmpty(str) && str.Length > 1)
-            {
-                return char.ToLowerInvariant(str[0]) + str.Substring(1);
-            }
-            return str.ToLowerInvariant();
-        }
-    }
-}
